feat: clean up code and properly add connection pool
